rfc2136: rename to dnsupdate
